backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

pydantic[email]==2.10.4
pydantic-settings==2.7.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

redis==5.2.1
arq==0.26.1

qdrant-client==1.12.1

google-cloud-aiplatform==1.74.0

opencv-python-headless==4.10.0.84
ffmpeg-python==0.2.0
numpy==2.2.1
pillow==11.1.0

httpx==0.28.1
